+++ Libs/Helper/PluginHelper.php
@@ -36,7 +36,7 @@ class PluginHelper extends \WordPress\Plugin\EveOnlineIntelTool\Libs\Singletons\
 	 * @return string
 	 */
 	public function getPluginPath($file = '') {
-		return \WP_PLUGIN_DIR . '/' . \dirname(\dirname(\dirname(\plugin_basename(__FILE__)))) . '/' . $file;
+		return \WP_PLUGIN_DIR . '/' . $this->getPluginDirName() . '/' . $file;
 	} // public function getPluginPath($file = '')
 
 	/**
@@ -46,8 +46,17 @@ class PluginHelper extends \WordPress\Plugin\EveOnlineIntelTool\Libs\Singletons\
 	 * @return string
 	 */
 	public function getPluginUri($file = '') {
-		return \WP_PLUGIN_URL . '/' . \dirname(\dirname(\dirname(\plugin_basename(__FILE__)))) . '/' . $file;
+		return \WP_PLUGIN_URL . '/' . $this->getPluginDirName() . '/' . $file;
 	} // public function getPluginUri()
+
+	/**
+	 * Get the plugins directory base name
+	 *
+	 * @return string
+	 */
+	public function getPluginDirName() {
+		return \dirname(\dirname(\dirname(\plugin_basename(__FILE__))));
+	} // END public function getPluginDirName()
 
 	/**
 	 * Returning the plugins default settings